+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.8.0-rc.10",
-  genVersion: "2.879.11",
-  userAgent: "speakeasy-sdk/typescript 4.8.0-rc.10 2.879.11 2.0 @orq-ai/node",
+  sdkVersion: "4.8.0-rc.11",
+  genVersion: "2.879.13",
+  userAgent: "speakeasy-sdk/typescript 4.8.0-rc.11 2.879.13 2.0 @orq-ai/node",
 } as const;