+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-13T13:54:04.819Z",
+    "2026-04-14T06:07:18.726Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {