+++ packages/orq-rc/RELEASES.md
@@ -8259,3 +8259,13 @@ Based on:
 - [typescript v4.8.0-rc.10] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.10 - packages/orq-rc
+
+## 2026-04-14 06:10:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.4 (2.879.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.11] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.11 - packages/orq-rc